refactor: Loading Spinner CSS로 대체
추가로 로딩 스피너를 common 컴포넌트로 분리한다. 로딩스피너는 재사용성이 높으므로 상황에 따라 크기를 변경 용이하게 prop으로 받는다.
