+++ src/components/@common/LoadingSpinner.tsx
@@ -1,0 +1,7 @@
+import * as Styled from './LoadingSpinner.styles';
+
+const LoadingSpinner = () => {
+  return <Styled.Spinner size={'60px'} />;
+};
+
+export default LoadingSpinner;
