+++ src/types/index.d.ts
@@ -2,8 +2,3 @@ declare module '*.svg' {
   const content: string;
   export default content;
 }
-
-declare module '*.png' {
-  const content: string;
-  export default content;
-}
